+++ python/cudf/cudf/_lib/strings/replace_re.pyx
@@ -20,14 +20,16 @@ from libcpp.string cimport string
 
 def replace_re(Column source_strings,
                object pattern,
-               DeviceScalar repl,
+               object py_repl,
                size_type n):
     """
     Returns a Column after replacing occurrences regular
-    expressions `pattern` with `repl` in `source_strings`.
+    expressions `pattern` with `py_repl` in `source_strings`.
     `n` indicates the number of resplacements to be made from
     start. (-1 indicates all)
     """
+
+    cdef DeviceScalar repl = py_repl.device_value
 
     cdef unique_ptr[column] c_result
     cdef column_view source_view = source_strings.view()